+++ Setup Your Mac/Setup-Your-Mac-via-Dialog.bash
@@ -9,11 +9,16 @@
 #
 # HISTORY
 #
+<<<<<<< HEAD
 #   Version 1.4.0, 21-Nov-2022, Dan K. Snelson (@dan-snelson)
 #   https://snelson.us/2022/11/setup-your-mac-via-swiftdialog-1-3-1/
 #   - Signficantly enchanced Completion Action options
 #   - Improved Debug Mode behavior
 #   - Miscellaneous Improvements
+=======
+#   Version 1.5.0, 28-Nov-2022, Dan K. Snelson (@dan-snelson)
+#   - Prompt user for additional fields at Welcome dialog
+>>>>>>> v1.5.0-rc
 #
 ####################################################################################################
 
@@ -29,12 +34,17 @@
 # Script Version, Jamf Pro Script Parameters and default Exit Code
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
+<<<<<<< HEAD
 scriptVersion="1.4.0"
+=======
+scriptVersion="1.5.0"
+>>>>>>> v1.5.0-rc
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin/
-debugMode="${4:-"true"}"                # [ true (default) | false ]
-assetTagCapture="${5:-"false"}"         # [ true | false (default) ]
-completionActionOption="${6:-"wait"}"   # [ wait (default) | sleep (with seconds) | Shut Down | Shut Down Attended | Shut Down Confirm | Restart | Restart Attended | Restart Confirm | Log Out | Log Out Attended | Log Out Confirm ]
-scriptLog="${7:-"/var/tmp/org.churchofjesuschrist.log"}"
+scriptLog="${4:-"/var/tmp/org.churchofjesuschrist.log"}"
+debugMode="${5:-"true"}"                           # [ true (default) | false ]
+welcomeDialog="${6:-"true"}"                       # [ true (default) | false ]
+completionActionOption="${7:-"Restart Attended"}"  # [ wait | sleep (with seconds) | Shut Down | Shut Down Attended | Shut Down Confirm | Restart | Restart Attended (default) | Restart Confirm | Log Out | Log Out Attended | Log Out Confirm ]
+reconOptions=""                                    # Initialize dynamic recon options; built based on user's input at Welcome dialog
 exitCode="0"
 
 
@@ -43,7 +53,7 @@ exitCode="0"
 # Reflect Debug Mode in `infotext` (i.e., bottom, left-hand corner of each dialog)
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-if [[ ${debugMode} == "true" ]]; then
+if [[ "${debugMode}" == "true" ]]; then
     scriptVersion="DEBUG MODE | Dialog: v$(dialog --version) • Setup Your Mac: v${scriptVersion}"
 fi
 
@@ -56,7 +66,6 @@ fi
 dialogApp="/usr/local/bin/dialog"
 welcomeCommandFile=$( mktemp /var/tmp/dialogWelcome.XXX )
 setupYourMacCommandFile=$( mktemp /var/tmp/dialogSetupYourMac.XXX )
-setupYourMacPolicyArrayIconPrefixUrl="https://ics.services.jamfcloud.com/icon/hash_"
 failureCommandFile=$( mktemp /var/tmp/dialogFailure.XXX )
 jamfBinary="/usr/local/bin/jamf"
 loggedInUser=$( echo "show State:/Users/ConsoleUser" | scutil | awk '/Name :/ { print $3 }' )
@@ -65,8 +74,168 @@ loggedInUserFirstname=$( echo "$loggedInUserFullname" | cut -d " " -f 1 )
 
 
 
+####################################################################################################
+#
+# Welcome dialog
+#
+####################################################################################################
+
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# APPS TO BE INSTALLED (Thanks, Obi-@smithjw!)
+# "Welcome" dialog Title, Message and Icon
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+welcomeTitle="Welcome to your new Mac, ${loggedInUserFirstname}!"
+welcomeMessage="To begin, please enter the required information below, then click **Continue** to start applying settings to your new Mac.  \n\nOnce completed, the **Quit** button will be re-enabled and you'll be prompted to restart your Mac.  \n\nIf you need assistance, please contact the Help Desk: +1 (801) 555-1212."
+
+# Welcome icon set to either light or dark, based on user's Apperance setting (thanks, @mm2270!) 
+appleInterfaceStyle=$( /usr/bin/defaults read /Users/"${loggedInUser}"/Library/Preferences/.GlobalPreferences.plist AppleInterfaceStyle 2>&1 )
+if [[ "${appleInterfaceStyle}" == "Dark" ]]; then
+    welcomeIcon="https://cdn-icons-png.flaticon.com/512/740/740878.png"
+else
+    welcomeIcon="https://cdn-icons-png.flaticon.com/512/979/979585.png"
+fi
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# "Welcome" JSON (thanks, @bartreardon!)
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+welcomeJSON='{
+    "title" : "'"${welcomeTitle}"'",
+    "message" : "'"${welcomeMessage}"'",
+    "icon" : "'"${welcomeIcon}"'",
+    "iconsize" : "198.0",
+    "button1text" : "Continue",
+    "button2text" : "Quit",
+    "infotext" : "'"${scriptVersion}"'",
+    "blurscreen" : "true",
+    "ontop" : "true",
+    "titlefont" : "size=26",
+    "messagefont" : "size=16",
+    "textfield" : [
+        {   "title" : "Comment",
+            "required" : false,
+            "prompt" : "Enter a comment",
+            "editor" : true
+        },
+        {   "title" : "Computer Name",
+            "required" : false,
+            "prompt" : "Computer Name"
+        },
+        {   "title" : "User Name",
+            "required" : false,
+            "prompt" : "User Name"
+        },
+        {   "title" : "Asset Tag",
+            "required" : true,
+            "prompt" : "Please enter the seven-digit Asset Tag",
+            "regex" : "^(AP|IP)?[0-9]{7,}$",
+            "regexerror" : "Please enter (at least) seven digits for the Asset Tag, optionally preceed by either AP or IP."
+        }
+    ],
+  "selectitems" : [
+        {   "title" : "Department",
+            "default" : "Please select your department",
+            "values" : [
+                "Please select your department",
+                "Asset Management",
+                "Australia Area Office",
+                "Board of Directors",
+                "Business Development",
+                "Corporate Communications",
+                "Creative Services",
+                "Customer Service / Customer Experience",
+                "Engineering",
+                "Finance / Accounting",
+                "General Management",
+                "Human Resources",
+                "Information Technology / Technology",
+                "Investor Relations",
+                "Legal",
+                "Marketing",
+                "Operations",
+                "Product Management",
+                "Production",
+                "Project Management Office",
+                "Purchasing / Sourcing",
+                "Quality Assurance",
+                "Risk Management",
+                "Sales",
+                "Strategic Initiatives & Programs",
+                "Technology"
+            ]
+        },
+        {   "title" : "Select B",
+            "values" : [
+                "B1",
+                "B2",
+                "B3"
+            ]
+        },
+        {   "title" : "Select C",
+            "values" : [
+                "C1",
+                "C2",
+                "C3"
+            ]
+        }
+    ],
+    "height" : "635"
+}'
+
+
+
+####################################################################################################
+#
+# Setup Your Mac dialog
+#
+####################################################################################################
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# "Setup Your Mac" dialog Title, Message, Overlay Icon and Icon
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+title="Setting up ${loggedInUserFirstname}'s Mac"
+message="Please wait while the following apps are installed …"
+overlayicon=$( defaults read /Library/Preferences/com.jamfsoftware.jamf.plist self_service_app_path )
+
+# Set initial icon based on whether the Mac is a desktop or laptop
+if system_profiler SPPowerDataType | grep -q "Battery Power"; then
+    icon="SF=laptopcomputer.and.arrow.down,weight=semibold,colour1=#ef9d51,colour2=#ef7951"
+else
+    icon="SF=desktopcomputer.and.arrow.down,weight=semibold,colour1=#ef9d51,colour2=#ef7951"
+fi
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# "Setup Your Mac" dialog Settings and Features
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+dialogSetupYourMacCMD="$dialogApp \
+--title \"$title\" \
+--message \"$message\" \
+--icon \"$icon\" \
+--progress \
+--progresstext \"Initializing configuration …\" \
+--button1text \"Wait\" \
+--button1disabled \
+--infotext \"$scriptVersion\" \
+--titlefont 'size=28' \
+--messagefont 'size=14' \
+--height '70%' \
+--position 'centre' \
+--blurscreen \
+--ontop \
+--overlayicon \"$overlayicon\" \
+--quitkey k \
+--commandfile \"$setupYourMacCommandFile\" "
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# "Setup Your Mac" policies to execute (Thanks, Obi-@smithjw!)
 #
 # For each configuration step, specify:
 # - listitem: The text to be displayed in the list
@@ -76,9 +245,13 @@ loggedInUserFirstname=$( echo "$loggedInUserFullname" | cut -d " " -f 1 )
 # - trigger: The Jamf Pro Policy Custom Event Name
 # - path: The filepath for validation
 #
-# shellcheck disable=1112
-#
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+# The fully qualified domain name of the server which hosts your icons, including any required sub-directories
+# (P.S. I tried to come up with a longer variable name, but couldn't.)
+setupYourMacPolicyArrayIconPrefixUrl="https://ics.services.jamfcloud.com/icon/hash_"
+
+# shellcheck disable=SC1112 # use literal slanted single quotes for typographic reasons
 
 policy_array=('
 {
@@ -161,10 +334,6 @@ policy_array=('
                 {
                     "trigger": "reconAtReboot",
                     "path": ""
-                },
-                {
-                    "trigger": "computerNameSet",
-                    "path": ""
                 }
             ]
         },
@@ -185,6 +354,7 @@ policy_array=('
 
 
 
+<<<<<<< HEAD
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # "Welcome / Asset Tag" Dialog Title, Message and Icon
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
@@ -202,84 +372,26 @@ else
 fi
 
 
+=======
+####################################################################################################
+#
+# Failure dialog
+#
+####################################################################################################
+>>>>>>> v1.5.0-rc
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# "Welcome / Asset Tag" Dialog Settings and Features
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-
-dialogWelcomeCMD="$dialogApp \
---title \"$welcomeTitle\" \
---message \"$welcomeMessage\" \
---icon \"$welcomeIcon\" \
---iconsize 198 \
---button1text \"Continue\" \
---button2text \"Quit\" \
---button2disabled \
---infotext \"$scriptVersion\" \
---blurscreen \
---ontop \
---titlefont 'size=26' \
---messagefont 'size=16' \
---textfield \"Asset Tag\",required=true,prompt=\"Please enter your Mac's seven-digit Asset Tag\",regex='^(AP|IP)?[0-9]{7,}$',regexerror=\"Please enter (at least) seven digits for the Asset Tag, optionally preceed by either 'AP' or 'IP'. \" \
---quitkey k \
---commandfile \"$welcomeCommandFile\" "
-
-
-
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# "Setup Your Mac" Dialog Title, Message, Overlay Icon and Icon
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-
-title="Setting up ${loggedInUserFirstname}'s Mac"
-message="Please wait while the following apps are installed …"
-overlayicon=$( defaults read /Library/Preferences/com.jamfsoftware.jamf.plist self_service_app_path )
-
-# Set initial icon based on whether the Mac is a desktop or laptop
-if system_profiler SPPowerDataType | grep -q "Battery Power"; then
-  icon="SF=laptopcomputer.and.arrow.down,weight=semibold,colour1=#ef9d51,colour2=#ef7951"
-else
-  icon="SF=desktopcomputer.and.arrow.down,weight=semibold,colour1=#ef9d51,colour2=#ef7951"
-fi
-
-
-
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# "Setup Your Mac" Dialog Settings and Features
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-
-dialogSetupYourMacCMD="$dialogApp \
---title \"$title\" \
---message \"$message\" \
---icon \"$icon\" \
---progress \
---progresstext \"Initializing configuration …\" \
---button1text \"Wait\" \
---button1disabled \
---infotext \"$scriptVersion\" \
---titlefont 'size=28' \
---messagefont 'size=14' \
---height '70%' \
---position 'centre' \
---blurscreen \
---ontop \
---overlayicon \"$overlayicon\" \
---quitkey k \
---commandfile \"$setupYourMacCommandFile\" "
-
-
-
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# "Failure" Dialog Title, Message and Icon
+# "Failure" dialog Title, Message and Icon
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 failureTitle="Failure Detected"
-failureMessage="Placeholder message; update in the finalise function"
+failureMessage="Placeholder message; update in the 'finalise' function"
 failureIcon="SF=xmark.circle.fill,weight=bold,colour1=#BB1717,colour2=#F31F1F"
 
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# "Failure" Dialog Settings and Features
+# "Failure" dialog Settings and Features
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 dialogFailureCMD="$dialogApp \
@@ -300,7 +412,8 @@ dialogFailureCMD="$dialogApp \
 
 
 
-#------------------------------- Edits below this line are optional -------------------------------#
+#------------------------ With the execption of the `finalise` function, -------------------------#
+#------------------------ edits below these line are optional. -----------------------------------#
 
 
 
@@ -377,6 +490,7 @@ function updateScriptLog() {
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # Run command as logged-in user (thanks, @scriptingosx!)
+# shellcheck disable=SC2145
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function runAsUser() {
@@ -419,9 +533,12 @@ function dialogCheck() {
         if [[ "$expectedDialogTeamID" == "$teamID" ]]; then
 
             /usr/sbin/installer -pkg "$tempDirectory/Dialog.pkg" -target /
+            sleep 2
+            updateScriptLog "swiftDialog version $(dialog --version) installed; proceeding..."
 
         else
 
+            # Display a so-called "simple" dialog if Team ID fails to validate
             runAsUser osascript -e 'display dialog "Please advise your Support Representative of the following error:\r\r• Dialog Team ID verification failed\r\r" with title "Setup Your Mac: Error" buttons {"Close"} with icon caution'
             completionActionOption="Quit"
             exitCode="1"
@@ -443,7 +560,7 @@ function dialogCheck() {
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Execute a "Welcome / Asset Tag" Dialog command
+# Update the "Welcome" dialog
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function dialogUpdateWelcome(){
@@ -454,7 +571,7 @@ function dialogUpdateWelcome(){
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Execute a "Setup Your Mac" Dialog command
+# Update the "Setup Your Mac" dialog
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function dialogUpdateSetupYourMac() {
@@ -465,7 +582,7 @@ function dialogUpdateSetupYourMac() {
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Execute a "Failure" Dialog command
+# Update the "Failure" dialog
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function dialogUpdateFailure(){
@@ -476,7 +593,7 @@ function dialogUpdateFailure(){
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Finalise app installations
+# Finalise User Experience
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function finalise(){
@@ -485,6 +602,7 @@ function finalise(){
 
         killProcess "caffeinate"
         updateScriptLog "Jamf Pro Policy Name Failures: ${jamfProPolicyPolicyNameFailures}"
+        dialogUpdateSetupYourMac "title: Sorry ${loggedInUserFirstname}, something went sideways"
         dialogUpdateSetupYourMac "icon: SF=xmark.circle.fill,weight=bold,colour1=#BB1717,colour2=#F31F1F"
         dialogUpdateSetupYourMac "progresstext: Failures detected. Please click Continue for troubleshooting information."
         dialogUpdateSetupYourMac "button1text: Continue …"
@@ -509,6 +627,7 @@ function finalise(){
 
     else
 
+        dialogUpdateSetupYourMac "title: ${loggedInUserFirstname}'s Mac is ready!"
         dialogUpdateSetupYourMac "icon: SF=checkmark.circle.fill,weight=bold,colour1=#00ff44,colour2=#075c1e"
         dialogUpdateSetupYourMac "progresstext: Complete! Please ${progressTextCompletionAction}enjoy your new Mac, ${loggedInUserFirstname}!"
         dialogUpdateSetupYourMac "progress: complete"
@@ -542,27 +661,47 @@ function get_json_value() {
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# smithjw's sweet function to execute Jamf Pro Policy Custom Events
+# Parse JSON via osascript and JavaScript for the Welcome dialog (thanks, @bartreardon!)
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+function get_json_value_welcomeDialog () {
+    for var in "${@:2}"; do jsonkey="${jsonkey}['${var}']"; done
+    JSON="$1" osascript -l 'JavaScript' \
+        -e 'const env = $.NSProcessInfo.processInfo.environment.objectForKey("JSON").js' \
+        -e "JSON.parse(env)$jsonkey"
+}
+
+
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Execute Jamf Pro Policy Custom Events (thanks, @smithjw)
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 function run_jamf_trigger() {
+
     trigger="$1"
-    if [[ "$debugMode" == "true" ]]; then
-        updateScriptLog "SETUP YOUR MAC DIALOG: DEBUG MODE: $jamfBinary policy -event $trigger"
-        sleep 1
-    elif [[ "$trigger" == "recon" ]]; then
-        dialogUpdateSetupYourMac "listitem: index: $i, status: wait, statustext: Updating …, "
-        if [[ ${assetTagCapture} == "true" ]]; then
-            updateScriptLog "SETUP YOUR MAC DIALOG: RUNNING: $jamfBinary recon -assetTag ${assetTag}"
-            "$jamfBinary" recon -assetTag "${assetTag}"
-        else
-            updateScriptLog "SETUP YOUR MAC DIALOG: RUNNING: $jamfBinary recon"
-            "$jamfBinary" recon
+
+    if [[ "${debugMode}" == "true" ]]; then
+
+        updateScriptLog "SETUP YOUR MAC DIALOG: DEBUG MODE: TRIGGER: $jamfBinary policy -event $trigger"
+        if [[ "$trigger" == "recon" ]]; then
+            updateScriptLog "SETUP YOUR MAC DIALOG: DEBUG MODE: RECON: $jamfBinary recon ${reconOptions}"
         fi
+        sleep 1
+
+    elif [[ "$trigger" == "recon" ]]; then
+
+        dialogUpdateSetupYourMac "listitem: index: $i, status: wait, statustext: Updating …, "
+        updateScriptLog "SETUP YOUR MAC DIALOG: Updating computer inventory with the following reconOptions: \"${reconOptions}\" …"
+        eval "${jamfBinary} recon ${reconOptions}"
+
     else
+
         updateScriptLog "SETUP YOUR MAC DIALOG: RUNNING: $jamfBinary policy -event $trigger"
         "$jamfBinary" policy -event "$trigger"
+
     fi
+
 }
 
 
@@ -593,7 +732,7 @@ function killProcess() {
 
 function completionAction() {
 
-    if [[ ${debugMode} == "true" ]]; then
+    if [[ "${debugMode}" == "true" ]]; then
 
         # If Debug Mode is enabled, ignore specified `completionActionOption`, display simple dialog box and exit
         runAsUser osascript -e 'display dialog "Setup Your Mac is operating in Debug Mode.\r\r• completionActionOption == '"'${completionActionOption}'"'\r\r" with title "Setup Your Mac: Debug Mode" buttons {"Close"} with icon note'
@@ -731,7 +870,13 @@ function quitScript() {
         rm "${failureCommandFile}"
     fi
 
-    # Check for user clicking "Quit" at Welcome screen
+    # Remove any default dialog file
+    if [[ -e /var/tmp/dialog.log ]]; then
+        updateScriptLog "Removing default dialog file …"
+        rm /var/tmp/dialog.log
+    fi
+
+    # Check for user clicking "Quit" at Welcome dialog
     if [[ "${welcomeReturnCode}" == "2" ]]; then
         exitCode="1"
         exit "${exitCode}"
@@ -790,7 +935,7 @@ fi
 # Logging preamble
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-if [[ ${debugMode} == "true" ]]; then
+if [[ "${debugMode}" == "true" ]]; then
     updateScriptLog "\n\n###\n# ${scriptVersion}\n###\n"
 else
     updateScriptLog "\n\n###\n# Setup Your Mac (${scriptVersion})\n###\n"
@@ -841,66 +986,160 @@ caffeinate -dimsu -w $$ &
 # If Debug Mode is enabled, replace `blurscreen` with `movable`
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-if [[ ${debugMode} == "true" ]]; then
-    dialogWelcomeCMD=${dialogWelcomeCMD//blurscreen/moveable}
+if [[ "${debugMode}" == "true" ]]; then
+    welcomeJSON=${welcomeJSON//blurscreen/moveable}
     dialogSetupYourMacCMD=${dialogSetupYourMacCMD//blurscreen/moveable}
 fi
 
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Display Welcome Screen and capture user's interaction
+# Write Welcome JSON to disk
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-if [[ ${assetTagCapture} == "true" ]]; then
+echo "$welcomeJSON" > "$welcomeCommandFile"
 
-    assetTag=$( eval "$dialogWelcomeCMD" | awk -F " : " '{print $NF}' )
-    # dialogWelcomeProcessID=$!
 
-    if [[ -z "${assetTag}" ]]; then
+
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+# Display Welcome dialog and capture user's input
+# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+if [[ "${welcomeDialog}" == "true" ]]; then
+
+    welcomeResults=$( ${dialogApp} --jsonfile "$welcomeCommandFile" --json )
+    if [[ -z "${welcomeResults}" ]]; then
         welcomeReturnCode="2"
     else
         welcomeReturnCode="0"
     fi
 
-fi
-
-
-
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Evaluate User Interaction at Welcome Screen
-# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-
-if [[ "${assetTagCapture}" == "true" ]]; then
-
     case "${welcomeReturnCode}" in
 
-        0)  ## Process exit code 0 scenario here
-            updateScriptLog "WELCOME DIALOG: ${loggedInUser} entered an Asset Tag of ${assetTag} and clicked Continue"
+        0)  # Process exit code 0 scenario here
+            updateScriptLog "WELCOME DIALOG: ${loggedInUser} entered information and clicked Continue"
+
+            ###
+            # Extract the various values from the welcomeResults JSON
+            ###
+
+            comment=$(get_json_value_welcomeDialog "$welcomeResults" "Comment")
+            computerName=$(get_json_value_welcomeDialog "$welcomeResults" "Computer Name")
+            userName=$(get_json_value_welcomeDialog "$welcomeResults" "User Name")
+            assetTag=$(get_json_value_welcomeDialog "$welcomeResults" "Asset Tag")
+            department=$(get_json_value_welcomeDialog "$welcomeResults" "Department" "selectedValue")
+            selectB=$(get_json_value_welcomeDialog "$welcomeResults" "Select B" "selectedValue")
+            selectC=$(get_json_value_welcomeDialog "$welcomeResults" "Select C" "selectedValue")
+
+
+
+            ###
+            # Output the various values from the welcomeResults JSON to the log file
+            ###
+
+            updateScriptLog "WELCOME DIALOG: • Comment: $comment"
+            updateScriptLog "WELCOME DIALOG: • Computer Name: $computerName"
+            updateScriptLog "WELCOME DIALOG: • User Name: $userName"
+            updateScriptLog "WELCOME DIALOG: • Asset Tag: $assetTag"
+            updateScriptLog "WELCOME DIALOG: • Department: $department"
+            updateScriptLog "WELCOME DIALOG: • Select B: $selectB"
+            updateScriptLog "WELCOME DIALOG: • Select C: $selectC"
+
+
+
+            ###
+            # Evaluate Various User Input
+            ###
+
+            # Computer Name
+            if [[ -n "${computerName}" ]]; then
+
+                # UNTESTED, UNSUPPORTED "YOYO" EXAMPLE
+                updateScriptLog "WELCOME DIALOG: Set Computer Name …"
+                currentComputerName=$( scutil --get ComputerName )
+                currentLocalHostName=$( scutil --get LocalHostName )
+
+                # Sets LocalHostName to a maximum of 15 characters, comprised of first eight characters of the computer's
+                # serial number and the last six characters of the client's MAC address
+                firstEightSerialNumber=$( system_profiler SPHardwareDataType | awk '/Serial\ Number\ \(system\)/ {print $NF}' | cut -c 1-8 )
+                lastSixMAC=$( ifconfig en0 | awk '/ether/ {print $2}' | sed 's/://g' | cut -c 7-12 )
+                newLocalHostName=${firstEightSerialNumber}-${lastSixMAC}
+
+                if [[ "${debugMode}" == "true" ]]; then
+
+                    updateScriptLog "WELCOME DIALOG: DEBUG MODE: Renamed computer from: \"${currentComputerName}\" to \"${computerName}\" "
+                    updateScriptLog "WELCOME DIALOG: DEBUG MODE: Renamed LocalHostName from: \"${currentLocalHostName}\" to \"${newLocalHostName}\" "
+
+                else
+
+                    # Set the Computer Name to the user-entered value
+                    scutil --set ComputerName "${computerName}"
+
+                    # Set the LocalHostName to `newLocalHostName`
+                    scutil --set LocalHostName "${newLocalHostName}"
+
+                    # Delay required to reflect change …
+                    # … side-effect is a delay in the "Setup Your Mac" dialog appearing
+                    sleep 5
+                    updateScriptLog "WELCOME DIALOG: Renamed computer from: \"${currentComputerName}\" to \"$( scutil --get ComputerName )\" "
+                    updateScriptLog "WELCOME DIALOG: Renamed LocalHostName from: \"${currentLocalHostName}\" to \"$( scutil --get LocalHostName )\" "
+
+                fi
+
+            else
+
+                updateScriptLog "WELCOME DIALOG: ${loggedInUser} did NOT specify a new computer name"
+                updateScriptLog "WELCOME DIALOG: • Current Computer Name: \"$( scutil --get ComputerName )\" "
+                updateScriptLog "WELCOME DIALOG: • Current Local Host Name: \"$( scutil --get LocalHostName )\" "
+
+            fi
+
+            # User Name
+            if [[ -n "${userName}" ]]; then
+                # UNTESTED, UNSUPPORTED "YOYO" EXAMPLE
+                reconOptions+="-endUsername \"${userName}\" "
+            fi
+
+            # Asset Tag
+            if [[ -n "${assetTag}" ]]; then
+                reconOptions+="-assetTag \"${assetTag}\" "
+            fi
+
+            # Department
+            if [[ -n "${department}" ]]; then
+                # UNTESTED, UNSUPPORTED "YOYO" EXAMPLE
+                reconOptions+="-department \"${department}\" "
+            fi
+
+            # Output `recon` options to log
+            updateScriptLog "WELCOME DIALOG: reconOptions: ${reconOptions}"
+
+            ###
+            # Display "Setup Your Mac" dialog (and capture Process ID)
+            ###
+
             eval "${dialogSetupYourMacCMD[*]}" & sleep 0.3
             dialogSetupYourMacProcessID=$!
-            dialogUpdateSetupYourMac "message: Asset Tag reported as \`${assetTag}\`. $message"
             ;;
 
-        2)  ## Process exit code 2 scenario here
-            updateScriptLog "WELCOME DIALOG: ${loggedInUser} clicked Quit when prompted to enter Asset Tag"
+        2)  # Process exit code 2 scenario here
+            updateScriptLog "WELCOME DIALOG: ${loggedInUser} clicked Quit at Welcome dialog"
             completionActionOption="Quit"
             quitScript "1"
             ;;
 
-        3)  ## Process exit code 3 scenario here
+        3)  # Process exit code 3 scenario here
             updateScriptLog "WELCOME DIALOG: ${loggedInUser} clicked infobutton"
             osascript -e "set Volume 3"
             afplay /System/Library/Sounds/Glass.aiff
             ;;
 
-        4)  ## Process exit code 4 scenario here
+        4)  # Process exit code 4 scenario here
             updateScriptLog "WELCOME DIALOG: ${loggedInUser} allowed timer to expire"
-            eval "${dialogSetupYourMacCMD[*]}" & sleep 0.3
-            dialogSetupYourMacProcessID=$!
+            quitScript "1"
             ;;
 
-        *)  ## Catch all processing
+        *)  # Catch all processing
             updateScriptLog "WELCOME DIALOG: Something else happened; Exit code: ${welcomeReturnCode}"
             quitScript "1"
             ;;
@@ -909,9 +1148,12 @@ if [[ "${assetTagCapture}" == "true" ]]; then
 
 else
 
+    ###
+    # Display "Setup Your Mac" dialog (and capture Process ID)
+    ###
+
     eval "${dialogSetupYourMacCMD[*]}" & sleep 0.3
     dialogSetupYourMacProcessID=$!
-    updateScriptLog "dialogSetupYourMacProcessID: ${dialogSetupYourMacProcessID}"
 
 fi
 
@@ -964,7 +1206,7 @@ dialogUpdateSetupYourMac "progress: $progress_index"
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Close Welcome Screen
+# Close Welcome dialog
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
 dialogUpdateWelcome "quit:"
@@ -1001,7 +1243,7 @@ for (( i=0; i<dialog_step_length; i++ )); do
             # If the path variable has a value, check if that path exists on disk
             if [[ -f "$path" ]]; then
                 updateScriptLog "SETUP YOUR MAC DIALOG: INFO: $path exists, moving on"
-                if [[ "$debugMode" == "true" ]]; then sleep 0.5; fi
+                if [[ "${debugMode}" == "true" ]]; then sleep 0.5; fi
             else
                 run_jamf_trigger "$trigger"
             fi