+++ Setup Your Mac/CHANGELOG.md
@@ -1,5 +1,28 @@
 # CHANGELOG
 
+<<<<<<< HEAD
+=======
+##  1.5.0
+### 28-Nov-2022
+[Release-specific Blog Post](https://snelson.us/2022/11/setup-your-mac-via-swiftdialog-1-5-0/)
+- :new: Prompt user for additional fields at Welcome dialog
+  - New fields are included in a single `welcomeJSON` variable (thanks for all your efforts and feedback, @drtaru and @Andrew!)
+    - See [welcomeScreenTesting.bash](JSON/welcomeScreenTesting.bash) (thanks, @bartreardon!)
+    - In **Debug Mode**, changes are logged only (thanks, @Andrew!)
+  - Dynamic `reconOptions` based on user's input at the **Welcome** dialog
+  - Thanks for your patience, @remusache, @midiman1000, @erikmadams, @colorenz and @benphilware
+- :fire: **Breaking Changes** :fire: (for users of **Setup Your Mac** prior to `1.5.0`)
+  - Script Parameter Reordering (sorry; I'll strive not to ever do this again)
+    - **Parameter 4:** Script Log Location [ `/var/tmp/org.churchofjesuschrist.log` ]
+    - **Parameter 5:** Debug Mode [ `true` (default) | `false` ]
+    - **Parameter 6:** Welcome dialog [ `true` (default) | `false` ]
+    - **Parameter 7:** Completion Action [ `wait` | `sleep` (with seconds) | `Shut Down` | `Shut Down Attended` | `Shut Down Confirm` | `Restart` |  **`Restart Attended`(default)** | `Restart Confirm` | `Log Out` | `Log Out Attended` | `Log Out Confirm` ]
+- Miscellaneous Improvements
+  - Moved code blocks and variables to better reflect the **Welcome** > **Setup Your Mac** > **Failure** workflow
+  - Random code clean-up
+
+
+>>>>>>> v1.5.0-rc
 ## 1.4.0
 ### 21-Nov-2022
 [Release-specific Blog Post](https://snelson.us/2022/11/setup-your-mac-via-swiftdialog-1-3-1)
@@ -32,7 +55,8 @@
 
 
 ## 1.3.0
-### 09-Nov-2022 
+### 09-Nov-2022
+[Release-specific Blog Post](https://snelson.us/2022/11/setup-your-mac-via-swiftdialog-1-3-0/)
 - **Script Parameter Changes:**
   - :warning: **Parameter 4:** `debug` mode **enabled** by default
   - :new: **Parameter 7:** Script Log Location
@@ -64,6 +88,7 @@
 - Additional tweaks discovered during internal production deployment
 ## 1.2.7
 ### 10-Sep-2022
+[Release-specific Blog Post](https://snelson.us/2022/09/setup-your-mac-via-swiftdialog-1-2-7/)
 - Added "completionAction" (Script Parameter 6) to address [Pull Request No. 5](https://github.com/dan-snelson/dialog-scripts/pull/5)
 - Added "Failure" dialog to address [Issue No. 6](https://github.com/dan-snelson/dialog-scripts/issues/6)
 ## 1.2.6
@@ -76,6 +101,7 @@
 ## 1.2.4
 
 ### 18-Aug-2022
+[Release-specific Blog Post](https://snelson.us/2022/06/setup-your-mac-via-swiftdialog-1-2-1/)
 - Swap "Installing …" and "Pending …" status indicators (thanks, @joncrain)
 ## 1.2.3
 ### 15-Aug-2022